Refactor settings views into separate files for better code organization
- Extract SimpleSettingsView to Views/Settings/SettingsView.swift
- Extract HelpSupportView to Views/Settings/HelpSupportView.swift
- Extract PrivacyPolicyView to Views/Settings/PrivacyPolicyView.swift
- Extract TermsOfServiceView to Views/Settings/TermsOfServiceView.swift
- Reduce TimerView.swift from 837 to 108 lines (-729 lines)
- Improve maintainability and separation of concerns
- Preserve all existing functionality and imports
